+++ .gitignore
@@ -21,3 +21,4 @@ Pipfile
 docs/
 
 uv.lock
+.venv/
