+++ packages/airless-google-cloud-core/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "airless-google-cloud-core"
-version = "0.0.5"
+version = "0.1.0.dev"
 description = "Airless package with the main components to use airless in google cloud"
 readme = "README.md"
 requires-python = ">=3.9"
