+++ packages/airless-google-cloud-core/requirements.txt
@@ -1,3 +1,3 @@
 google-cloud-pubsub>=2.13.11,<3.0.0
 
-airless-core>=0.1.1.dev6,<0.2.0
+airless-core>=0.2.1.dev1,<0.3.0
