+++ packages/airless-google-cloud-storage/airless/google/cloud/storage/hook/__init__.py
@@ -1,4 +1,5 @@
-from .storage import (GcsHook, GcsDatalakeHook)
+from .storage import (GcsHook)
+from .datalake import (GcsDatalakeHook)
 
 __all__ = [
     'GcsHook',
