+++ requirements-dev.txt
@@ -1,5 +1,5 @@
 build==0.10.0
-twine==5.1.1
+twine==6.1.0
 bump2version==1.0.1
 pytest==8.1.1
 pytest-cov==4.1.0
